refactor: User -> Reservation 수정
