+++ ReservationService/src/main/resources/application.properties
@@ -1,6 +1,6 @@
-spring.application.name=UserService
+spring.application.name=ReservationService
 
-server.port = 8081
+server.port = 8087
 
 # DB connect
 spring.datasource.url=${SPRING_DATASOURCE_URL}
